+++ packages/dbsync/src/index.ts
@@ -1,3 +1,2 @@
-export * from "./DbRepository.js"
+export * from "./DbTable.js"
 export * from "./DbSync.js"
-export * from "./DbTxRepository.js"